modified

+++ FetchToDoListFunc/FetchToDoListFunc.cs
@@ -11,6 +11,7 @@ using FetchToDoListFunc.Repository;
 using System.Collections.Generic;
 using FetchToDoListFunc.Model;
 using MongoDB.Bson;
+using FetchToDoListFunc.Utility;
 
 namespace FetchToDoListFunc
 {
@@ -24,26 +25,42 @@ namespace FetchToDoListFunc
         }
 
         [FunctionName("Fetch")]
-        public static string Fetch(
-            [HttpTrigger(AuthorizationLevel.Function, "get", Route = "fetch")] HttpRequest req,
+        public static async Task<ActionResult> Fetch(
+            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "fetch")] HttpRequest req,
             ILogger log)
         {
             log.LogInformation("C# HTTP trigger function processed a request.");
 
-            return "Hurray";
+            // Check if we have authentication info.
+            ValidateJWT auth = new ValidateJWT(req);
+            if (!auth.IsValid)
+            {
+                return new UnauthorizedResult(); // No authentication info.
+            }
+            return new OkObjectResult("hurray"); ;
         }
 
         [FunctionName("getall")]
-        public async Task<ActionResult<IEnumerable<TaskList>>> GetAllTask([HttpTrigger(AuthorizationLevel.Function, "get", Route = null)] HttpRequest req, ILogger logger)
+        public async Task<ActionResult<IEnumerable<TaskList>>> GetAllTask([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)] HttpRequest req, ILogger logger)
         {
+            ValidateJWT auth = new ValidateJWT(req);
+            if (!auth.IsValid)
+            {
+                return new UnauthorizedResult(); // No authentication info.
+            }
             return new OkObjectResult(await _fetchRepo.GetAllAsync());
         }
 
         [FunctionName("add-task")]
-        public async Task<ActionResult<List<TaskList>>> CreateAsync([HttpTrigger(AuthorizationLevel.Function, "post", Route = null)] HttpRequest req, ILogger logger)
+        public async Task<ActionResult<List<TaskList>>> CreateAsync([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)] HttpRequest req, ILogger logger)
         {
             try
             {
+                ValidateJWT auth = new ValidateJWT(req);
+                if (!auth.IsValid)
+                {
+                    return new UnauthorizedResult(); // No authentication info.
+                }
                 var reqBody = await new StreamReader(req.Body).ReadToEndAsync();
                 var input = JsonConvert.DeserializeObject<TaskList>(reqBody);
                 var task = new TaskList
@@ -66,17 +83,26 @@ namespace FetchToDoListFunc
         }
 
         [FunctionName("gettaskbytaskname")]
-        public async Task<ActionResult<IEnumerable<TaskList>>> GetTaskDetails([HttpTrigger(AuthorizationLevel.Function, "get", Route = "gettaskbytaskname/{taskname}")] HttpRequest req, ILogger logger,string taskname)
+        public async Task<ActionResult<IEnumerable<TaskList>>> GetTaskDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "gettaskbytaskname/{taskname}")] HttpRequest req, ILogger logger,string taskname)
         {
+            ValidateJWT auth = new ValidateJWT(req);
+            if (!auth.IsValid)
+            {
+                return new UnauthorizedResult(); // No authentication info.
+            }
             return new OkObjectResult(await _fetchRepo.GetTaskDetailsByTaskNameAsync(taskname));
         }
 
         [FunctionName("update-task")]
-        public async Task<ActionResult<List<TaskList>>> UpdateTaskStatus([HttpTrigger(AuthorizationLevel.Function, "put", Route = "update-task/{id}")] HttpRequest req, ILogger logger,string id)
+        public async Task<ActionResult<List<TaskList>>> UpdateTaskStatus([HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "update-task/{id}")] HttpRequest req, ILogger logger,string id)
         {
             try
             {
-
+                ValidateJWT auth = new ValidateJWT(req);
+                if (!auth.IsValid)
+                {
+                    return new UnauthorizedResult(); // No authentication info.
+                }
                 return new OkObjectResult(await _fetchRepo.UpdateTaskStatusAsync(ObjectId.Parse(id)));
             }
             catch (Exception ex)
@@ -87,10 +113,15 @@ namespace FetchToDoListFunc
         }
 
         [FunctionName("delete-task")]
-        public async Task<ActionResult> DeleteProduct([HttpTrigger(AuthorizationLevel.Function, "delete", Route = "delete-product/{id}")] HttpRequest request, ILogger logger, string id)
+        public async Task<ActionResult> DeleteProduct([HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "delete-product/{id}")] HttpRequest req, ILogger logger, string id)
         {
             try
             {
+                ValidateJWT auth = new ValidateJWT(req);
+                if (!auth.IsValid)
+                {
+                    return new UnauthorizedResult(); // No authentication info.
+                }
                 return new OkObjectResult(await _fetchRepo.DeleteTaskAsync(ObjectId.Parse(id)));
             }
             catch (Exception ex)
